feat: Add pretrained ViT model (#61)
* feat: Add ViT model

#60

* style: modify a word from label to task

#60
