+++ config.json
@@ -3,7 +3,7 @@
     "n_gpu": 1,
 
     "arch": {
-        "type": "MultiLabelEfficientB0",
+        "type": "ViT",
         "args": {}
     },
     "dataset": {
